+++ Arduino/proyecto_final_bebederos/slave_esp32/water_level_sensor.cpp
@@ -1,8 +1,6 @@
 #include "water_level_sensor.h"
 
-//int water_pin = 33;  // Define el número del pin D15 (GPIO15)
-//int water_pin = 34;
-uint8_t value = 0;
+int value = 0;
 
 void set_water_level_parameters(){
   pinMode(PIN_SIGNAL, INPUT);
@@ -11,7 +9,7 @@ void set_water_level_parameters(){
 }
 
 
-uint8_t read_water_level() {
+int read_water_level() {
   digitalWrite(POWER,HIGH);
   delay(10);
   value = analogRead(PIN_SIGNAL);
